trade against BTC

+++ Main.py
@@ -231,7 +231,7 @@ def calc_stake_amount(coin):
             telegram.send_telegram_message(telegramToken, telegram.eWarning, msg)
     
         tradable_balance = balance*tradable_balance_ratio 
-        stake_amount = int(tradable_balance/(max_number_of_open_positions-num_open_positions))
+        stake_amount = tradable_balance/(max_number_of_open_positions-num_open_positions)
         
         # make sure the size is >= the minimum size
         if stake_amount < min_position_size:
